Moved files to ensure better file structure
