+++ src/components/atoms/ContactBadgeComponent.js
@@ -6,19 +6,19 @@ import React from 'react'
 /**
  * Styles
  */
-import styles from '../styles/ContactBadgeComponent.module.css'
+import styles from '../../styles/ContactBadgeComponent.module.css'
 
-const ContactBadgeComponent = (props) => (
+const ContactBadgeComponent = () => (
     <div className={styles.contactBadgeComponentContainer}>
         <span className={styles.badgeText}>
             Neem direct contact op
         </span>
         <div className={styles.badgeContainer}>
             <a href='/contact'>
-                <img src={'/images/bubble.svg'} />
+                <img src={'/images/bubble.svg'} alt={'Chat Bubble Icon'}/>
             </a>
         </div>
-        <img className={styles.badgeArrow} src={'/images/arrow.svg'} />
+        <img className={styles.badgeArrow} src={'/images/arrow.svg'} alt={'Arrow icon'} />
     </div>
 )
