+++ src/styles/atoms/BookComponent.module.css
